FAQ Column Widths
I changed the widths of the keyboard shortcuts in FAQ page. Let me know
if I should change it anymore. I did not want to put the What it does
column further to the right because then you would have to move your
eyes more to read the “What it does”. I was thinking, maybe replacing
“What it does” with “Function” or something like that. But either one
works.